refactor(models): move the MRF and fvOptions contributions next to their specs
The solver's pressure_velocity/extension.py now only defines the sites;
each model's contributions live at the bottom of neofoam/mrf.py /
neofoam/fv_options.py, below the spec they decorate — the solver package
imports those modules back to register the specs, so the extension
import must come after the spec exists.

Byte-parity re-verified on mixerVessel2D (MRF) and simpleCar (fvOptions).
